+++ Runtime/Helper/NetworkUtility.cs
@@ -132,7 +132,7 @@ namespace JFramework.Net
         /// <param name="action">传入网络连接，网络消息，传输通道</param>
         /// <typeparam name="T">网络消息</typeparam>
         /// <returns>返回一个消息委托</returns>
-        internal static MessageDelegate GetMessage<T>(Action<NetworkClient, T, int> action) where T : struct, Message
+        internal static MessageDelegate GetMessage<T>(Action<NetworkClient, T, byte> action) where T : struct, Message
         {
             return (client, reader, channel) =>
             {
